Moved Interfaces to seperate dir
