+++ ModelDefinition/Column/Decorators/Nullable.cls.php
@@ -2,7 +2,7 @@
 
 namespace ModelDefinition\Column\Decorators;
 
-use ModelDefinition\Column\ColumnInterface;
+use ModelDefinition\Column\Interfaces\ColumnInterface;
 
 final class Nullable implements ColumnInterface
 {
